+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Added validation for vertex weights being normalized.
 * Added support for loading folders by dragging and dropping files or folders onto the application.
 * Added a list view tab to the Anim Editor for showing animation track values.
+* Added validation for premultiplied shaders with a Source Color of SourceAlpha.
 
 ### Changed
 * Adjusted the behavior of adding folders and reloading folders to preserve toggled model and mesh visibility.